+++ src/main/java/org/bboxdb/tools/gui/ChooseDistributionGroupDialog.java
@@ -31,6 +31,7 @@ import javax.swing.JFrame;
 import org.bboxdb.distribution.DistributionGroupName;
 import org.bboxdb.distribution.zookeeper.ZookeeperClient;
 import org.bboxdb.distribution.zookeeper.ZookeeperException;
+import org.bboxdb.distribution.zookeeper.ZookeeperNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -201,7 +202,7 @@ public class ChooseDistributionGroupDialog {
 							while(! scalepahntGUI.shutdown) {
 								try {
 									guiModel.updateDistributionRegion();
-								} catch (ZookeeperException e) {
+								} catch (ZookeeperException | ZookeeperNotFoundException e) {
 									logger.warn("Got exception: ", e);
 								}
 								scalepahntGUI.updateView();