+++ changelog.md
@@ -7,6 +7,7 @@
 - Improvement: Added CLI parameter to the SSTableExaminer
 - Improvement: The documentation is now pushed to 'GitHub pages'
 - Improvement: Rewrote the KD-Tree zookeeper integration code
+- Improvement: Improved handling of deleted nodes in zookeeper
 - Bugfix: Init table splitter only in distributed mode
 - Bugfix: In a major compact, not all tables were processed.